+++ Lang/Backend/AssemblyBackend.cs
@@ -2,7 +2,7 @@ namespace Lang.Backend.Assembly
 {
     public class AssemblyBackend : IBackend
     {
-        public void Build(ProjectFile project, ProgramGraph programGraph, BuildSettings buildSettings)
+        public string Build(ProjectFile project, ProgramGraph programGraph, BuildSettings buildSettings)
         {
             throw new System.NotImplementedException();
         }
