+++ Lang/Backend/IBackend.cs
@@ -8,6 +8,6 @@ namespace Lang.Backend
         /// <param name="projectPath">The path to the project</param>
         /// <param name="programGraph">Graph of the program</param>
         /// <param name="buildSettings">Build settings from the cli args</param>
-        void Build(ProjectFile project, ProgramGraph programGraph, BuildSettings buildSettings);
+        string Build(ProjectFile project, ProgramGraph programGraph, BuildSettings buildSettings);
     }
 }
