+++ Lang/Program.cs
@@ -10,6 +10,7 @@ serviceCollection.AddTransient<IProjectInterpreter, ProjectInterpreter>();
 serviceCollection.AddTransient<IProgramGraphBuilder, ProgramGraphBuilder>();
 serviceCollection.AddTransient<IPolymorpher, Polymorpher>();
 serviceCollection.AddTransient<IProgramRunner, ProgramRunner>();
+serviceCollection.AddTransient<ILinker, Linker>();
 // LLVM Backend
 serviceCollection.AddTransient<IBackend, LLVMBackend>();
